Added listener for animations to start and end.
AnimationListener added for listening to animation starts and stops,
StreamDeckController listens to those animation starts and stops,
displaying the default icon as soon as the animation ends.
AnimationStack now defines if an animation should be ended as soon as
the trigger is not valid anymore or if the animation should finish.
*Listener moved to de.rcblum.stream.deck.items.listeners.