Added a software based stream deck interface
Additional to the stream deck hardware, the library can now be used
without it(the hardware). This helps developers that want to develop for
Stream Deck but don't have one yet. The software interface can be
disabled by setting StreamDeckDevices.ENABLE_SOFTWARE_STREAM_DECK to
false at the start of the program.