No duplicate Animators for a key on a IStreamDeck more possible
When there was an Error on the stream deck's hid interface, it was
possible for the Animators to not be removed properly. This was now
fixed in the Animator class by caching the animators for specific stream
decks and key numbers and disabling them, when a new animator is set for
the key.